LayoutTests/platform/chromium/fast/forms/week/week-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for  from renderer. No cases of empty initial values for type=week.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
FAIL stepUp("2010-W02", null, null) should be 2010-W03. Was 2010-W02.
FAIL stepDown("2010-W02", null, null) should be 2010-W01. Was 2010-W02.
FAIL stepUp("2010-W02", null, null, 10) should be 2010-W12. Was 2010-W02.
FAIL stepDown("2010-W02", null, null, 11) should be 2009-W44. Was 2010-W02.
FAIL stepUp("1970-W01", "4", null, 2) should be 1970-W09. Was 1970-W01.
FAIL stepDown("1970-W01", "4", null, 3) should be 1969-W41. Was 1970-W01.
Step=any
FAIL stepUp("2010-W02", "any", null) should be 2010-W03. Was 2010-W02.
FAIL stepDown("2010-W02", "any", null) should be 2010-W01. Was 2010-W02.
Overflow/underflow
FAIL stepUp("2010-W02", "3.40282346e+38", null) should be 275760-W37. Was 2010-W02.
FAIL stepDown("2010-W02", "3.40282346e+38", null) should be 1970-W01. Was 2010-W02.
PASS stepUp("2010-W02", "1", "2010-W02") is "2010-W02"
PASS stepDown("2010-W02", "1", "2010-W02") is "2010-W02"
stepDown()/stepUp() for stepMismatch values
FAIL stepDown("2010-W02", "2", "2009-W52") should be 2010-W01. Was 2010-W02.
FAIL stepUp("1970-W02", "4", "") should be 1970-W05. Was 1970-W02.

PASS successfullyParsed is true

TEST COMPLETE
